+++ utils/scarb-oracle-hint-processor/Cargo.toml
@@ -14,9 +14,6 @@ repository.workspace = true
 
 [dependencies]
 anyhow.workspace = true
-cairo-lang-casm.workspace = true
-cairo-lang-runner.workspace = true
-cairo-vm.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 shell-words.workspace = true
