+++ utils/scarb-oracle-hint-processor/src/connection.rs
@@ -1,12 +1,12 @@
 use crate::connections::stdio_jsonrpc::StdioJsonRpcConnection;
 use anyhow::{Result, bail};
-use cairo_vm::Felt252;
+use starknet_core::types::Felt;
 use std::collections::HashMap;
 use std::collections::hash_map::Entry;
 use std::fmt;
 
 pub trait Connection {
-    fn call(&mut self, selector: &str, calldata: &[Felt252]) -> Result<Vec<Felt252>>;
+    fn call(&mut self, selector: &str, calldata: &[Felt]) -> Result<Vec<Felt>>;
 }
 
 /// Maintains a collection of oracle [`Connection`]s.
